Revert to the old alsps driver form 3.10.61
Some old chips doesn't work with the driver in the 3.10.72 source. Don't
know why Lenovo change it in the first place.
A piece of code would be an art to me. I spent 20 minutes to tidy up the
codes and unifying the indents. It is much easier to read now. This is
not OCD. It would be time saving for later or other's review.
-daniel_hk